requirements.txt
Flask>=3.0
requests>=2.31
feedparser>=6.0.11
beautifulsoup4>=4.12
lxml>=5.2
readability-lxml>=0.8.1   # optional fallback; install it
